Feature : Big Decimal realisation

+++ app/src/main/java/com/infinity_coder/divcalendar/data/db/dao/SecurityDao.kt
@@ -2,6 +2,7 @@ package com.infinity_coder.divcalendar.data.db.dao
 
 import androidx.room.*
 import com.infinity_coder.divcalendar.data.db.model.SecurityDbModel
+import java.math.BigDecimal
 
 @Dao
 abstract class SecurityDao {
@@ -18,7 +19,7 @@ abstract class SecurityDao {
     abstract suspend fun getSecurityPackagesForPortfolio(portfolioId: Long): List<SecurityDbModel>
 
     @Query("SELECT count(*) FROM ${SecurityDbModel.TABLE_NAME} WHERE ${SecurityDbModel.COLUMN_PORTFOLIO_ID} = :portfolioId")
-    abstract suspend fun getSecurityCountForPortfolio(portfolioId: Long): Int
+    abstract suspend fun getSecurityCountForPortfolio(portfolioId: Long): BigDecimal
 
     @Delete
     abstract suspend fun deleteSecurityPackage(securityPackage: SecurityDbModel)
