+++ app/src/main/java/com/infinity_coder/divcalendar/presentation/portfolio/PortfolioFragment.kt
@@ -26,6 +26,7 @@ import com.infinity_coder.divcalendar.presentation.portfolio.sorting.SortingPort
 import com.infinity_coder.divcalendar.presentation.search.SearchSecurityActivity
 import kotlinx.android.synthetic.main.fragment_portfolio.*
 import kotlinx.android.synthetic.main.layout_stub_empty.view.*
+import java.math.BigDecimal
 
 class PortfolioFragment : Fragment(R.layout.fragment_portfolio),
     ChangeSecurityBottomDialog.OnClickListener,
@@ -169,7 +170,7 @@ class PortfolioFragment : Fragment(R.layout.fragment_portfolio),
         }
     }
 
-    private fun setTotalPortfolioCost(totalPortfolioCost: Double) {
+    private fun setTotalPortfolioCost(totalPortfolioCost: BigDecimal) {
         val currentCurrency = viewModel.getDisplayCurrency()
         val totalPortfolioCostWithCurrency = SecurityCurrencyDelegate.getValueWithCurrencyConsiderCopecks(requireContext(), totalPortfolioCost, currentCurrency)
         totalPortfolioCostTextView.text = getString(R.string.total_portfolio_cost, totalPortfolioCostWithCurrency)
